requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.3
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.10
django-allauth==0.61.1
django-bootstrap-v5==1.0.11
django-bootstrap5==23.4
django-heroku==0.3.1
gunicorn==21.2.0
idna==3.6
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3
whitenoise==6.6.0
django-storages
boto3
selenium==3.141.0
responses==0.21.0